feat: add loading screen phases, LOD-first meshing, persistent settings, and UI polish
Bootstrap & content loading:
- Convert ContentPipeline.Build() to IEnumerable<string> iterator that
  yields phase descriptions between processing stages
- Move bootstrap to Start() coroutine so loading screen is visible
  during content pipeline phases via LoadingScreen.SetContentPhase()
- Cache resolved block models in ContentModelResolver to avoid
  redundant parent-chain traversal

LOD scheduling:
- Split LODScheduler back into UpdateLODLevels + ScheduleJobs so LOD
  level assignment runs before MeshScheduler decides who to mesh
- LODScheduler now handles Generated chunks with LODLevel > 0
  (first-time LOD mesh), not just Ready chunk LOD transitions
- Add ChunkManager.FillGeneratedChunks and FillGeneratedChunksWithLOD
  query methods for the new scheduling flow
- Extract AssignLODLevels, IsLODPending, ScheduleLODMesh helpers
- Add spawn radius vs LOD1 distance assertion in bootstrap

UI:
- HotbarHUD shows item name tooltip above hotbar on slot change with
  timed fade-out; FormatFullItemName converts snake_case to Title Case
- InventoryScreen dirty-checks slot state to skip unchanged slot updates
- SettingsScreen persists render distance, FOV, mouse sensitivity, and
  AO strength to PlayerPrefs; loads and applies on Initialize

Other:
- Change NativeNoiseConfig.SeedOffset from int to float for fractional
  density noise offsets
- Fix load queue distance sort to use float instead of int for lengthsq
- Add LootResolver determinism tests (same-seed, weighted entries)
- PhysicsSettings: walkSpeed 4.3→12, gravity -28→0 (debug/testing values)

+++ Packages/com.lithforge.worldgen/Runtime/Noise/NativeNoiseConfig.cs
@@ -10,6 +10,6 @@ namespace Lithforge.WorldGen.Noise
         public float Persistence;
         public float HeightScale;
         public int Octaves;
-        public int SeedOffset;
+        public float SeedOffset;
     }
 }
